JDK 1.4 compatibility fixes.

+++ main/java/org/mvel/conversion/BigDecimalCH.java
@@ -44,7 +44,7 @@ public class BigDecimalCH implements ConversionHandler {
     }
 
     static {
-        CNV.put(Object.class,
+         CNV.put(Object.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
                         return new BigDecimal(String.valueOf(o));
@@ -64,7 +64,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(BigInteger.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((BigInteger) o);
+                        return new BigDecimal(((BigInteger) o).doubleValue());
                     }
                 }
         );
@@ -80,7 +80,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(Double.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((Double) o);
+                        return new BigDecimal(((Double) o).doubleValue());
                     }
                 }
         );
@@ -88,7 +88,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(Float.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((Float) o);
+                        return new BigDecimal(((Float) o).doubleValue());
                     }
                 }
         );
@@ -97,7 +97,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(Short.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((Short) o);
+                        return new BigDecimal(((Short) o).doubleValue());
                     }
                 }
         );
@@ -105,7 +105,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(Long.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((Long) o);
+                        return new BigDecimal(((Long) o).doubleValue());
                     }
                 }
         );
@@ -113,7 +113,7 @@ public class BigDecimalCH implements ConversionHandler {
         CNV.put(Integer.class,
                 new Converter() {
                     public BigDecimal convert(Object o) {
-                        return new BigDecimal((Integer) o);
+                        return new BigDecimal(((Integer) o).doubleValue());
                     }
                 }
         );
